requirements.txt
fastapi
uvicorn
python-telegram-bot
langchain
langchain-community
lanchain_core
ollama
python-dotenv
httpx
duckduckgo-search